app/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.0
django-environ==0.4.5
django-js-asset==1.2.2
djangorestframework==3.12.4
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
